requirements.txt
Django==4.1.5
python-dotenv==0.21.0
django-allauth~=0.52.0
django-summernote~=0.8.20.0
django-filter~=22.1
celery~=5.2.7
redis~=4.4.0